refactor(LottoPrize) : float형 double형으로 변경 및 적용

+++ src/main/java/lotto/controller/LottoController.java
@@ -56,7 +56,7 @@ public class LottoController {
         outputView.printStatistics(lottoStatistics);
     }
 
-    private void printEarning(float earning){
+    private void printEarning(double earning){
         outputView.printEarning(earning);
     }
