+++ src/main/java/lotto/domain/LottoPrize.java
@@ -4,10 +4,10 @@ import java.util.Map;
 
 public class LottoPrize {
 
-    public float calculateEarning(int lottoPrice) {
-        return ((float) addPrize() / (float) lottoPrice) * 100;
+    public double calculateEarning(int lottoPrice) {
+        return (double) addPrize() / lottoPrice * 100;
     }
-    
+
     private int addPrize(){
         Map<LottoResult,Integer>  lottoResultMap = LottoStatistics.getLottoResultMap();
         int total = lottoResultMap.keySet()
